test(project): refactor gateway add coverage
